Refactor tests abstracting mutability
